financial statement type guard
